Addressed almost all SonarLint issues - Addressed all small issues - The only remaining issues are giant complexity ones, for example in main.Main, ui.UserInterface, or ui.DialogHelper - Hopefully this doesn't introduce more :)

+++ src/main/java/ui/export/TempFiles.java
@@ -29,9 +29,8 @@ class TempFiles implements AutoCloseable {
 
 	@Override
 	public void close() {
-		if (isTempMonthFile) {
-			if (!monthFile.delete())
-				monthFile.deleteOnExit();
+		if (isTempMonthFile && !monthFile.delete()) {
+			monthFile.deleteOnExit();
 		}
 	}
